While previous pr #3736 improved both memory/throughput for aggregation, binary, and unary PromQL queries by loading only necessary columns into memory. Other queries still load the entire parquet file, resulting in high memory usage and slow throughput. 

This pr breaks the data loading process into two steps by utilizing `__hash__` columns. 
1. find the unique combinations of `__hash__` and all k-v pairs of labels
2. load only `__hash__`, `_timestamp`, and `value` columns into memory for querying

This will greatly improve the performance for vector/matrix selector queries. For aggregation, binary, and unary queries, memory footprint will be further reduced while enduring small throughput impact. 

Performance comparison of an aggregation query`max by (node) (container_memory_usage_bytes)` raised by #3091 

Before:
```
time range | max_mem_usage | fetching ms | processing ms | total ms
1 days     | 569mb         | 115         | 60            | 175
2 days     | 852mb         | 217         | 120           | 337
7 days     | 2.90g         | 750         | 473           | 1223
2 weeks    | 4.5g          | 1530        | 843           | 2373
1 month    | 8.9g          | 3241        | 1723          | 4964
```

after refactor:
```
time range  | max_mem_usage | step1  | process1 ms | step2 ms | process2 ms | total ms
1 days      | 470mb         | 267    | 0           | 59       | 59          | 385
2 days      | 642mb         | 428    | 1           | 115      | 123         | 667
7 days      | 2.16g         | 1410   | 4           | 521      | 470         | 2405
2 weeks     | 3.26g         | 2233   | 6           | 1070     | 848         | 4157
1 month     | 6.6g          | 4788   | 12          | 2482     | 1841        | 9123
```

PromQL compliance result
```
================================================================================
General query tweaks:
*  Openobserve is sometimes off by 1ms when parsing floating point start/end timestamps.
*  Openobserve fractional tolerance amount for <agg>_over_time queries
================================================================================
Total: 538 / 548 (98.18%) passed, 0 unsupported
```