@hengfeiyang 2024-06-20
@taimingl i checked this PR, You run the `before` with the PR only select a `node` field, but run the `after` with all fields, in the `after` we also can apply the change select the fields only we need.